Fix opts(legend.position={left,top,bottom}).
When trying to change the legend position to left, top, or bottom using
opts(), one ends up with a broken graph.  This bug can be located in
surround_viewports() where the graphical elements are placed into the
grid.  If we increment all rows and columns of the graphical elements for
positions "left", "top", and "bottom" by 1, those graphs look sane again.
I assume that a new first row and column were added at some point in the
development (commit db7474fb ?), but only the parameters for the default
position "right" and for position "none" were adjusted.