Allow density to expand scale bounds
In the previous implementation, density was estimated only within the
range of the scale, which is determined by default by the range of the
data. So, to see the effect of not trimming the data, the user had to
explicitly expand the scale.

In the new implementation, with trim = FALSE, the density is indeed
estimated up to 3 bandwidths after the bounds of the data, and this
expands the scale. But trim = TRUE by default (as was said in the
documentation but not respected in the code) so the final aspect should
stay the same in most circumstances.