First stub of a length scale
It can be used whenever fixed segment lengths are drawn. Examples are given
with geom_rug but it could also be used with a segment geom that would be
defined in terms of angle and length rather than in terms of beginning and end
coords.

The remaining issue is the length, which currently just draws segments. It
should draw segments of lengths identical to the main plot but since the two
viewports (main plot and legend) do not have the same reference, it might be
hard.