gg_dep fixes: argument name and multiple arguments passed
The documentation for gg_dep and the function arguments disagreed on
the name of the second argument. Fixed documentation to agree with code.

Two instances of multiple character strings being passed to gg_dep
(where it was possible to pass them to message/warning) were fixed
by wrapping them in paste with a sep of \n (like was done in other
functions).