Improved graphic for shapes catalog
Revamped the graphic for showing the different shapes (0-25) utilizing
facet_wrap instead of staggering along the axes. Also using theme to
get rid of unnecessary elements, but retainging the grid lines that show
exactly where the points are centered. Changed the fill from black to
lightblue and made the symbols bigger.