Rewrite of helper functions for dynamic Rd files
Removed geom_aesthetics and stat_aesthetics since they were
not used anywhere else in the codebase

Refactored aesthetics and rd_aesthetics so that aesthetics only
collects the information about a given geom or stat's aesthetics
and rd_aesthetics is responsible for all the formatting of this
into Rd format. Parameters of the two functions are the same, but
aesthetics now returns a data.frame rather than a character vector.

rd_aesthetics now also outputs the default values for aesthetics.

roxygen-like documentation added for aesthetics function (same format
as roxygen documentation, but with just comment prefix rather than
roxygen prefix) as a type of internal documentation.