Fix for last small segment in polar coordinate munching being dropped
When in polar coordinates, when the last segment to be drawn
was short enough that it did not require any interpolation,
the segment was not drawn because the munch_data code did not
include the last (second) point of the segment.

Better calculation of how many segments the segment should
be interpolated into.

Put in a missing drop=FALSE when subsetting a data.frame because
there may be only a single column (which would then not be
assembled with the rest of the data properly).

Fixed interp so that the code matches the documentation. Also
clarified the documentation.

Wrote test code to exercise interp and munch_data