Add hook to modify plot_data prior to transform
This will make it possible to pass a function as the data argument that will modify the global data for the layer. It is indeed an esoteric use case, but for ggplot2 extension developers it is the only way to intercept the global data prior to passing it through the relevant Stat object.

Furthermore the change is really minor and cheap. If you dislike the approach please consider another way for extension builders to modify the data prior to Stat...