Expose argument n from stats::density in the interface of stat_density. (#1661)
* Expose argument n from stats::density in the interface of stat_density.

The argument controls the number of equally spaced points at which the
density is to be estimated. As in stats::density the argument n defaults
to 512. It is exposed through the function interface of stat_density and
can thus also be used from geom_density.

An existent local variable n, which holds the number of values in
the input vector, has been renamed to nx to avoid a name clash.

* Added a note about powers of two being recommended.