Added `expand_scale()` function for easier generation of scale expansion vectors.
Instead of having to manually specify an `expand` argument using
a somewhat confusing syntax (a vector of 2, 3 or 4 numeric values),
it’s now possible to use the user-friendly (and documented)
`expand_scale()` function.

This commit also cleans up the documentation related to the
`expand` argument, which was duplicated in several functions.