Allow separate expansion values for lower and upper range limits (v2) (#1805)
* Allow separate expansion values for lower and upper range limits.

The `expand` argument for `scale_*_continuous()` and `scale_*_discrete()`
now accepts separate expansion constants for the lower and upper range limits.

This is useful for creating bar charts where the bottom of the bars
are flush with the x axis but the bars still have some (automatically
calculated amount of) space above them:

```R
ggplot(mtcars) +
  geom_bar(aes(x = factor(cyl))) +
  scale_y_continuous(expand = c(0, 0, 0.1, 0))
```

It can also be useful for line charts, e.g. for counts over time,
where one wants to have a ’hard’ lower limit of y = 0, but leave the
upper limit unspecified (and perhaps differing between panels),
but with some extra space above the highest point on the line.
(With symmetrical limits, the extra space above the highest point
could cause the lower limit to be negative.)

The syntax for the multiplicative and additive expansion
constants has been changed from `c(m, a)` to
`c(m_lower, a_lower, m_uppper, a_upper)`. The old syntax will still
work, as length 2 vectors `c(m, a)` are expanded to `c(m, a, m, a)`
and length 3 vectors are expanded from `c(m1, a1, m2)` to
`c(m1, a2, m2, a1)`. (@huftis, #1669)

* Added `expand_scale()` function for easier generation of scale expansion
vectors.

Instead of having to manually specify an `expand` argument using
a somewhat confusing syntax (a vector of 2, 3 or 4 numeric values),
it’s now possible to use the user-friendly (and documented)
`expand_scale()` function.

This commit also cleans up the documentation related to the
`expand` argument, which was duplicated in several functions.

* Added UTF-8 character encoding declaration to DESCRIPTION.

The documentation for one of the functions had a no-breaking space,
(between a number and the word ‘units’), which caused R CMD check to
complain about ‘non-ASCII input and no declared encoding’.
This adds a character encoding declaration of UTF-8 to the DESCRIPTION
file to fix this problem.

* Fixed some style issues.

* Updated and regenerated documentation.

* Specify character encoding used for documentation.

* Minor grammar improvement in documentation.

* Don’t generate documentation for internal function expand_range4().