New examples for theme_*()
Update the example used in the "complete themes" help page (`ggtheme`). I extended the current `mtcars` example to use more (all?) of the features which themes affect: title, subtitle, tag, caption, and both x- and y-facets.

There are now two plot examples. One with no facets, and one with 2x2 facets. The latter is enclosed in `\donttest{}` to keep R CMD check from complaining about the runtime.

The `mtcars` dataset may not be the best for this, but I wanted to keep it close to the orginal example, and the data isn't as important in this example.