Add new parameter "preset" for ggsave() to specify width/height at once
Presets can be used to specify plot width and height (and units) at
once. It recognizes common paper sizes (such as "a4" and "letter"),
common screen resolutions ("4k"), or a resolution specification such
as "1920x1080". All presets default to landscape orientation, and
portrait can be specified by appending an "r" (for rotated) to the end
of the string. This is similar to the "paper" option for
grDevices::pdf(), but with landscape being the default as it's
probably a more common choice for plots.