Add facet variants called subplots
Facets are meant to display different subsets of the same data.

However, with data in the "long" format (i.e. several variables stacked
in the same column and an additional column indicating which is which),
they can also be used to create the same plot for several variables
(e.g. a histogram). In that case, the scales are often free (the
variables are different and hence often do not have the same unit).

The problem with that approach is that the axis title is often a generic,
uninformative word (e.g. "values") while the real labels for the axis
are the facet labels, which are on the wrong side of the plot.

These functions use `facet_wrap` and `facet_grid` to create subplots and
adjust the position and aspect of the facet labels to make them look like
axis titles.