hjust and vjust in geom_text() with angle %% 90 > 45
In geom_text() when abs(angle) %% 90  > 45 justification applied based on the orientation of the text labels is done along x axis for vjust and along y axis for hjust. For "outward" and "inward" to behave correctly, the position tested against 0.5 needs to change from x to y and vice versa based on the value of angle for each row in data.