Clean the code
This commit reverts some of the edits in the three previous commits.
Set signature of compute_just() as suggested by thomasp85. As angle is set in default aes to zero, we can assume angle will be always available: remove corresponding test. Set default arguments for new parameters of compute_just() so that geom_label() does not need edits and so that existing test cases remain valid. Add new test cases.