FEAT: ScaleContinuous can return native colour format
Some geoms benefit from using native colour format instead
of the character based colour format.

This commit lets the geoms specify that they prefer native
format for colours, and gives the responsibility of converting
into that format to ScaleContinuous.

Since today it is not mandatory for all scales to honor scale_params,
the geom that requests this will have to verify that the
color is given in native format anyway, and do the conversion if
it has not been done here.

However by optionally shifting the responsibility of the conversion
to the scale we have potential to further optimizations