PERF: Palette capability: accepts_native_output=TRUE
If a palette has accepts_native_output=TRUE set as an attribute,
ScaleContinuous assumes the palette has an optional argument named
`color_fmt` which can be set to either "character" or "native".

If the geom prefers a native output format and the palette supports
it, we let the palette take care of it.

The conversion goes from value -> native colour, which is much faster
than going through an intermediate character representation of
the colours.